Move HiveSchema to hive_ce
