+++ hive_ce_inspector/lib/connect_client.dart
@@ -1,13 +1,9 @@
-// This is internal access
-// ignore_for_file: implementation_imports
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:hive_ce/hive.dart';
-import 'package:hive_ce/src/connect/hive_connect_api.dart';
-import 'package:hive_ce/src/binary/binary_reader_impl.dart';
+import 'package:hive_ce_inspector/hive_internal.dart';
 import 'package:vm_service/vm_service.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
