+++ hive_ce_inspector/lib/hive_internal.dart
@@ -1,0 +1,2 @@
+export 'package:hive_ce/src/connect/hive_connect_api.dart';
+export 'package:hive_ce/src/binary/binary_reader_impl.dart';
