+++ hive_generator/lib/src/helper/helper.dart
@@ -5,8 +5,10 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:collection/collection.dart';
 import 'package:hive_ce/hive.dart';
+import 'package:hive_ce_generator/src/model/hive_schema.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:path/path.dart' as path;
+import 'package:yaml_writer/yaml_writer.dart';
 
 final _hiveFieldChecker = const TypeChecker.fromRuntime(HiveField);
 
@@ -96,4 +98,12 @@ extension BuildStepExtension on BuildStep {
       ..createSync(recursive: true)
       ..writeAsStringSync(content);
   }
+}
+
+/// Write a [HiveSchema] to a string
+String writeSchema(HiveSchema schema) {
+  final yaml = YamlWriter().write(schema.toJson());
+  return '''
+${HiveSchema.comment}
+$yaml''';
 }
