+++ hive_generator/pubspec.yaml
@@ -25,3 +25,7 @@ dev_dependencies:
   build_runner: any
   rexios_lints: ^9.1.0
   json_serializable: ^6.8.0
+
+dependency_overrides:
+  hive_ce:
+    path: ../hive
